+++ src/services/ros/index.ts
@@ -99,11 +99,6 @@ export enum AccessLevel {
   admin,
 }
 
-export interface IRealmSizeInfo {
-  stateSize?: number;
-  fileSize?: number;
-}
-
 // An enum to describe a users role
 
 export enum UserRole {