+++ src/ui/ServerAdministration/TopBar/Status.tsx
@@ -19,51 +19,82 @@
 import * as React from 'react';
 import * as Realm from 'realm';
 
-import { ILoadingProgress } from '../../reusable/LoadingOverlay';
+import { ILoadingProgress, LoadingStatus } from '../../reusable/LoadingOverlay';
+import { RealmAdminConnection } from '../AdminRealm';
+
+import { ConnectionStateIndicator } from './ConnectionStateIndicator';
+import { VersionBadge } from './VersionBadge';
 
 export interface IStatusProps {
   onReconnect: () => void;
   progress: ILoadingProgress;
   user: Realm.Sync.User | null;
+  serverVersion?: string;
 }
 
-export const Status = ({ onReconnect, progress, user }: IStatusProps) => {
-  if (user) {
-    if (progress.status === 'failed') {
-      return (
-        <p className="TopBar__Status">
-          <i className="fa fa-exclamation-circle" /> Disconnected: "
-          <span className="TopBar__Status__error">{progress.message}</span>
-          "&nbsp;
-          {/* progress.retry ? (
-            <Button size="sm" onClick={progress.retry.onRetry}>
-              Reconnect now
-            </Button>
-          ) : null */}
-        </p>
-      );
-    } else if (progress.status === 'in-progress') {
-      return (
-        <p className="TopBar__Status">
-          Connecting to&nbsp;
-          <span className="TopBar__Status__server">{user.server}</span>
-        </p>
-      );
-    } else if (progress.status === 'done') {
-      return (
-        <p className="TopBar__Status">
-          Connected to&nbsp;
-          <span className="TopBar__Status__server">{user.server}</span>
-        </p>
-      );
-    } else {
-      return (
-        <p className="TopBar__Status">
-          <i className="fa fa-exclamation-circle" /> Not connected
-        </p>
-      );
-    }
+function renderText(prefix: string, selectable: string = '') {
+  return (
+    <React.Fragment>
+      {prefix}
+      <span className="TopBar__Status__selectable">{selectable}</span>
+    </React.Fragment>
+  );
+}
+
+function renderStatus(
+  connectionState: Realm.Sync.ConnectionState,
+  status: LoadingStatus,
+  message?: string,
+  serverUrl?: string,
+) {
+  if (connectionState === Realm.Sync.ConnectionState.Connected) {
+    return renderText(serverUrl ? 'Connected to ' : 'Connected', serverUrl);
+  } else if (connectionState === Realm.Sync.ConnectionState.Connecting) {
+    return renderText(serverUrl ? 'Connecting to ' : 'Connecting', serverUrl);
+  } else if (
+    connectionState === Realm.Sync.ConnectionState.Disconnected &&
+    status === 'done'
+  ) {
+    return renderText(
+      serverUrl ? 'Disconnected from ' : 'Disconnected',
+      serverUrl,
+    );
+  } else if (status === 'failed') {
+    // We could include the message, but that is usually very long and its shown on the LoadingOverlay anyway
+    return 'Failed';
+  } else if (status === 'in-progress') {
+    return message || 'Loading';
+  } else if (status === 'done') {
+    return message || 'Authenticated';
   } else {
-    return <p className="TopBar__Status">Authenticating ...</p>;
+    return message || 'Waiting';
   }
+}
+
+export const Status = ({
+  progress: { status, message },
+  user,
+  serverVersion,
+}: IStatusProps) => {
+  return (
+    <RealmAdminConnection>
+      {connectionState => (
+        <React.Fragment>
+          <p className="TopBar__Status">
+            {renderStatus(
+              connectionState,
+              status,
+              message,
+              user ? user.server : undefined,
+            )}
+          </p>
+          <VersionBadge serverVersion={serverVersion} />
+          <ConnectionStateIndicator
+            className="TopBar__Status__Icon"
+            connectionState={connectionState}
+          />
+        </React.Fragment>
+      )}
+    </RealmAdminConnection>
+  );
 };