+++ CHANGELOG.md
@@ -51,7 +51,7 @@ This version had a regression running on Windows and was therefore never release
 - When browsing a reference Realm a permission sidebar appears in the browser. ([#946](https://github.com/realm/realm-studio/pull/946))
 
 ## Fixed
-- Fixed missing padding in the Realm browsers sidebar when a schema was missing ([#1026](https://github.com/realm/realm-studio/pull/1026))
+- Fixed missing padding in the Realm browsers sidebar when a schema was missing. ([#1026](https://github.com/realm/realm-studio/pull/1026))
 
 ## Internals
 - Excluding the unpackaged directories when uploading to S3 (again). ([#1013](https://github.com/realm/realm-studio/pull/1013))